+++ app_runner/docs/changelog.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## \[Unreleased\]
 
+### Added
+
+- Register single file command: `bfabric-app-runner outputs register-single-file`
+
 ## \[0.0.7\] - 2024-11-22
 
 ### Fixed